test funcionales

+++ gdx-logic-bricks/src/main/java/com/indignado/logicbricks/core/actuators/InstanceEntityActuator.java
@@ -10,7 +10,7 @@ import com.indignado.logicbricks.core.EntityFactory;
 public class InstanceEntityActuator<T extends Entity> extends Actuator {
     public EntityFactory entityFactory;
     public Vector2 localPosition;
-    public float duration = 0; // 0 = lives forever
+    public short duration = 0; // 0 = lives forever
     public Type type;
 
     public enum Type {AddEntity, RemoveEntity}
